Renamed libraries and packages to tigl3
This allows to use tigl2 and tigl3 at the same time.

Addresses issue #233
