+++ CMakeLists.txt
@@ -10,9 +10,9 @@ if(NOT DEFINED CMAKE_BUILD_TYPE)
 endif()
 
 project (TIGL)
-set(TIGL_VERSION_MAJOR 2)
-set(TIGL_VERSION_MINOR 2)
-set(TIGL_VERSION_PATCH 0)
+set(TIGL_VERSION_MAJOR 3)
+set(TIGL_VERSION_MINOR 0)
+set(TIGL_VERSION_PATCH 0dev)
 set(TIGL_VERSION "${TIGL_VERSION_MAJOR}.${TIGL_VERSION_MINOR}.${TIGL_VERSION_PATCH}")
 
 set(TIGL_ADD_INCLUDE_PATH "" CACHE PATH "Additional include path for package search")
