--- bindings/java/src/de/dlr/sc/tigl/TiglBSpline.java
@@ -1,28 +1,0 @@
-/* 
-* Copyright (C) 2007-2013 German Aerospace Center (DLR/SC)
-*
-* Created: 2014-10-21 Martin Siggel <martin.siggel@dlr.de>
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-
-package de.dlr.sc.tigl;
-
-import java.util.ArrayList;
-
-public class TiglBSpline {
-    public ArrayList<Double> knots = new ArrayList<>();
-    public ArrayList<TiglPoint> controlPoints = new ArrayList<>();
-    public int degree = 0;
-    
-}
