+++ bindings/java/src/de/dlr/sc/tigl3/TiglSymmetryAxis.java
@@ -17,34 +17,36 @@
 */
 
 /* 
-* This file is automatically created from tigl.h on 2014-10-21.
+* This file is automatically created from tigl.h on 2017-05-26.
 * If you experience any bugs please contact the authors
 */
 
-package de.dlr.sc.tigl;
+package de.dlr.sc.tigl3;
 
 import java.util.ArrayList;
 
-public enum TiglStructureType {
-    UPPER_SHELL(0),
-    LOWER_SHELL(1),
-    INNER_STRUCTURE(2);
+public enum TiglSymmetryAxis {
+    TIGL_NO_SYMMETRY(0),
+    TIGL_X_Y_PLANE(1),
+    TIGL_X_Z_PLANE(2),
+    TIGL_Y_Z_PLANE(3);
 
-    private static ArrayList<TiglStructureType> codes = new ArrayList<>();
+    private static ArrayList<TiglSymmetryAxis> codes = new ArrayList<>();
 
     static {
-        codes.add(UPPER_SHELL);
-        codes.add(LOWER_SHELL);
-        codes.add(INNER_STRUCTURE);
+        codes.add(TIGL_NO_SYMMETRY);
+        codes.add(TIGL_X_Y_PLANE);
+        codes.add(TIGL_X_Z_PLANE);
+        codes.add(TIGL_Y_Z_PLANE);
     }
 
     private final int code;
 
-    private TiglStructureType(final int value) {
+    private TiglSymmetryAxis(final int value) {
          code = value;
     }
 
-    public static TiglStructureType getEnum(final int value) {
+    public static TiglSymmetryAxis getEnum(final int value) {
         return codes.get(Integer.valueOf(value));
     }
