+++ bindings/matlab/Makefile.in
@@ -2,4 +2,4 @@
 # It requires a MATLAB installation and MATLAB's mex command must be in PATH
 
 all:
-	mex tiglmatlab.c -I../../../include -L../../../@CMAKE_INSTALL_LIBDIR@ -lTIGL -output tigl_matlab
+	mex tiglmatlab.c -I../../../include/tigl3 -L../../../@CMAKE_INSTALL_LIBDIR@ -ltigl3 -output tigl_matlab
