+++ examples/CMakeLists.txt
@@ -5,13 +5,13 @@ add_subdirectory(java_demo)
 
 install(FILES 
     ${PROJECT_SOURCE_DIR}/tests/TestData/simpletest.cpacs.xml
-    DESTINATION share/doc/tigl/examples
+    DESTINATION share/doc/tigl3/examples
     COMPONENT docu
 )
 
 install(FILES 
     README.md
-    DESTINATION share/doc/tigl/examples
+    DESTINATION share/doc/tigl3/examples
     COMPONENT docu
     RENAME README.txt
 )
