+++ examples/c_demo/CMakeLists.txt
@@ -4,14 +4,14 @@
 include_directories(${PROJECT_SOURCE_DIR}/src/api ${TIXI_INCLUDE_DIRS})
 
 
-set(demo_LIBS TIGL)
+set(demo_LIBS tigl3)
 
 add_executable(c_demo c_demo.c)
 target_link_libraries(c_demo ${demo_LIBS})
 
 install(FILES 
     c_demo.c
-    DESTINATION share/doc/tigl/examples
+    DESTINATION share/doc/tigl3/examples
     COMPONENT docu
 )
 
@@ -19,14 +19,14 @@ if(WIN32)
   # install makefile.msvc
   install(FILES 
       makefile.msvc
-      DESTINATION share/doc/tigl/examples
+      DESTINATION share/doc/tigl3/examples
       COMPONENT docu
   )
 else(WIN32)
   # install makefile.gnu  
   install(FILES 
       makefile.gnu
-      DESTINATION share/doc/tigl/examples
+      DESTINATION share/doc/tigl3/examples
       COMPONENT docu
   )
 endif(WIN32)
