+++ examples/java_demo/CMakeLists.txt
@@ -3,7 +3,7 @@
 
 install(FILES 
     JavaDemo.java
-    DESTINATION share/doc/tigl/examples
+    DESTINATION share/doc/tigl3/examples
     COMPONENT docu
 )
