+++ examples/matlab_demo/CMakeLists.txt
@@ -3,7 +3,7 @@
 
 install(FILES 
     matlab_demo.m
-    DESTINATION share/doc/tigl/examples
+    DESTINATION share/doc/tigl3/examples
     COMPONENT docu
 )
