+++ examples/python_internal/example_step_export.py
@@ -1,8 +1,8 @@
 from __future__ import print_function
 
 from tixi import tixiwrapper
-from tigl import tiglwrapper
-import tigl.configuration, tigl.geometry, tigl.boolean_ops, tigl.exports
+from tigl3 import tigl3wrapper
+import tigl3.configuration, tigl3.geometry, tigl3.boolean_ops, tigl3.exports
 import os
 
 def export_fused_step(tigl_handle):
@@ -42,7 +42,7 @@ def export_fused_step(tigl_handle):
 
 if __name__ == '__main__':
     tixi_h = tixiwrapper.Tixi()
-    tigl_h = tiglwrapper.Tigl()
+    tigl_h = tigl3wrapper.Tigl3()
 
     dir_path = os.path.dirname(os.path.realpath(__file__))
     tixi_h.open(dir_path + "/../../tests/TestData/simpletest.cpacs.xml")
